+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+* Own page for privacy statement
+* Broken link checker to discover changed URLs
+
 ### Changed
 * Details: URLs based on the name of the church
 * Show recently added entries on home page instead of open data page
 * Don't mark HTTPS links with secure icon (only HTTP marked as non-secure)
-* Own page for privacy statement
 
 ### Fixed
 * Add Form: Fix geolocation